requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Werkzeug==2.3.7
pandas==2.1.1
openpyxl==3.1.2
selenium==4.15.2
webdriver-manager==4.0.1
googlemaps==4.10.0
python-dotenv==1.0.0
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4
Pillow==10.0.1
zipfile36==0.1.3
psycopg2-binary==2.9.7
mysqlclient==2.1.1
prometheus-client==0.17.1
structlog==23.1.0 